Associate overview viewmodel with view in App.xaml.cs
Remove the reference to the overview viewmodel in the shell viewmodel.
Instead, use the singleton locator to load the object (if it exists)
and change its properties based on wallet events.

This can be further cleaned up by having the overview viewmodel
subscribe directly with wallet events, rather than mutating it from
the shell viewmodel's event handlers.  To do this, the wallet object
potentially needs to be associated with the viewmodel at a later time
than construction, so this can happen as part of the wallet's synced
event handler.